requirements.txt
psycopg2-binary==2.8.6
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
flake8>=3.6.0,<3.7.0
psycopg2>=2.7.5,<2.8.0
gunicorn>=20.0.4,<20.1.0
Pillow>=5.3.0,<5.4.0